Added an ArchUnit test verifying that Application does not depend on Adapters and fixed a violation in test code by moving DummyNotifier into the application.ports package.
